requirements.txt
fastapi
uvicorn
gunicorn
pydantic
azure-storage-blob
pandas
numpy
tableauhyperapi